REFACTOR :: move security package into shard
